requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.3
